pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "promptix"
version = "0.1.14"
description = "A simple library for managing and using prompts locally with Promptix Studio"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Promptix Contributors", email = "contact@promptix.io"}
]
keywords = ["prompt", "management", "llm", "ai", "studio", "templates"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Text Processing :: General",
]
dependencies = [
    "streamlit>=1.29.0",
    "jinja2>=3.0.0",
    "python-dotenv>=0.19.0",
    "pyyaml>=6.0.0",
    "jsonschema>=4.0.0",
]

[project.scripts]
promptix = "promptix.tools.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "build>=1.0.0",
    "twine>=4.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
"promptix.tools.studio" = ["*.webp", "*.ico"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=promptix --cov-report=term-missing"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3 